pyproject.toml
[project]
name = "zerosyl"
version = "0.1.0"
description = "Simple zero-resource syllable tokenization for spoken language modeling"
authors = [
    {name = "nicolvisser",email = "vissernicol@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11,<3.14"
dependencies = [
    "torch (>=2.10.0,<3.0.0)",
    "torchcodec (>=0.10.0,<0.11.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "typer (>=0.23.0,<0.24.0)",
    "transformers (>=5.1.0,<6.0.0)",
    "tgt (>=1.5,<2.0)",
    "pandas (>=3.0.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "torchaudio (>=2.10.0,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "wandb (>=0.25.0,<0.26.0)"
]

[project.scripts]
zerosyl = "zerosyl.cli:app"